refactor: isolate backend layers
