+++ backend/query/users.sql
@@ -1,5 +1,5 @@
 -- name: user.selectByUsername
-SELECT id, username, name, email, role
+SELECT id, username, password_hash, name, email, role
 FROM T_USER
 WHERE username = :u;
 
@@ -8,4 +8,7 @@ SELECT COUNT(*) AS cnt FROM T_USER;
 
 -- name: sys.ping
 SELECT 1;
+
+-- name: sys.oraclePing
+SELECT 1 FROM DUAL;
