+++ backend/service/TransactionService.py
@@ -14,25 +14,20 @@ async def ensure_tables() -> None:
     if "main_db" not in DB.dbManagers:
         return
     db = DB.dbManagers["main_db"]
-    await db.execute("""
-    CREATE TABLE IF NOT EXISTS test_transaction (
-        id INTEGER PRIMARY KEY AUTOINCREMENT,
-        value TEXT UNIQUE
-    )
-    """)
+    await db.executeQuery("tx.ensureTable")
 
 
 async def do_single_commit() -> None:
     await ensure_tables()
     db = DB.dbManagers["main_db"]
-    await db.execute("INSERT INTO test_transaction (value) VALUES (:val)", {"val": "tx-single"})
+    await db.executeQuery("tx.insertValue", {"val": "tx-single"})
 
 
 async def do_unique_violation() -> None:
     await ensure_tables()
     db = DB.dbManagers["main_db"]
     # insert a fixed value to hit unique constraint
-    await db.execute("INSERT INTO test_transaction (value) VALUES (:val)", {"val": "tx-dup"})
+    await db.executeQuery("tx.insertValue", {"val": "tx-dup"})
     # second insert will violate UNIQUE
-    await db.execute("INSERT INTO test_transaction (value) VALUES (:val)", {"val": "tx-dup"})
+    await db.executeQuery("tx.insertValue", {"val": "tx-dup"})
